🐛 不小心把 main 的返回值也给改了。

+++ luogu/P1010.cpp
@@ -20,7 +20,7 @@
 
 using namespace std;
 
-long main()
+int main()
 {
     long x;
     cin >> x;
